Added: Float Test Case For 1st Position

+++ src/main/java/com/bdlz/generics/Maximum.java
@@ -2,8 +2,8 @@ package com.bdlz.generics;
 
 public class Maximum {
 
-        public static Integer findMaximum(Integer a, Integer b, Integer c) {
-            Integer max = a;
+        public static Float findMaximum(Float a, Float b, Float c) {
+            Float max = a;
             if (b.compareTo(a)>0 && b.compareTo(c)>0) {
                 max = b;
             } if (c.compareTo(max) > 0) {
